consolidate code, tests
