+++ pkg/core/actions/getmetrics.go
@@ -23,6 +23,7 @@ package actions
 
 import (
 	m "github.com/blackducksoftware/perceptor/pkg/core/model"
+	log "github.com/sirupsen/logrus"
 )
 
 // GetMetrics .....
@@ -32,6 +33,88 @@ type GetMetrics struct {
 
 // Apply .....
 func (g *GetMetrics) Apply(model *m.Model) {
-	modelMetrics := model.Metrics()
+	modelMetrics := metrics(model)
 	go g.Continuation(modelMetrics)
+}
+
+func metrics(model *m.Model) *m.Metrics {
+	// number of images in each status
+	statusCounts := make(map[m.ScanStatus]int)
+	for _, imageResults := range model.Images {
+		statusCounts[imageResults.ScanStatus]++
+	}
+
+	// number of containers per pod (as a histgram, but not a prometheus histogram ???)
+	containerCounts := make(map[int]int)
+	for _, pod := range model.Pods {
+		containerCounts[len(pod.Containers)]++
+	}
+
+	// number of times each image is referenced from a pod's container
+	imageCounts := make(map[m.Image]int)
+	for _, pod := range model.Pods {
+		for _, cont := range pod.Containers {
+			imageCounts[cont.Image]++
+		}
+	}
+	imageCountHistogram := make(map[int]int)
+	for _, count := range imageCounts {
+		imageCountHistogram[count]++
+	}
+
+	podStatus := map[string]int{}
+	podPolicyViolations := map[int]int{}
+	podVulnerabilities := map[int]int{}
+	for podName := range model.Pods {
+		podScan, err := scanResultsForPod(model, podName)
+		if err != nil {
+			log.Errorf("unable to get scan results for pod %s: %s", podName, err.Error())
+			continue
+		}
+		if podScan != nil {
+			podStatus[podScan.OverallStatus.String()]++
+			podPolicyViolations[podScan.PolicyViolations]++
+			podVulnerabilities[podScan.Vulnerabilities]++
+		} else {
+			podStatus["Unknown"]++
+			podPolicyViolations[-1]++
+			podVulnerabilities[-1]++
+		}
+	}
+
+	imageStatus := map[string]int{}
+	imagePolicyViolations := map[int]int{}
+	imageVulnerabilities := map[int]int{}
+	for sha, imageInfo := range model.Images {
+		if imageInfo.ScanStatus == m.ScanStatusComplete {
+			imageScan := imageInfo.ScanResults
+			if imageScan == nil {
+				log.Errorf("found nil scan results for completed image %s", sha)
+				continue
+			}
+			imageStatus[imageScan.OverallStatus().String()]++
+			imagePolicyViolations[imageScan.PolicyViolationCount()]++
+			imageVulnerabilities[imageScan.VulnerabilityCount()]++
+		} else {
+			imageStatus["Unknown"]++
+			imagePolicyViolations[-1]++
+			imageVulnerabilities[-1]++
+		}
+	}
+
+	// TODO
+	// number of images without a pod pointing to them
+	return &m.Metrics{
+		ScanStatusCounts:      statusCounts,
+		NumberOfImages:        len(model.Images),
+		NumberOfPods:          len(model.Pods),
+		ContainerCounts:       containerCounts,
+		ImageCountHistogram:   imageCountHistogram,
+		PodStatus:             podStatus,
+		ImageStatus:           imageStatus,
+		PodPolicyViolations:   podPolicyViolations,
+		ImagePolicyViolations: imagePolicyViolations,
+		PodVulnerabilities:    podVulnerabilities,
+		ImageVulnerabilities:  imageVulnerabilities,
+	}
 }
