+++ pkg/core/actions/getscanresults.go
@@ -22,8 +22,12 @@ under the License.
 package actions
 
 import (
+	"fmt"
+
 	"github.com/blackducksoftware/perceptor/pkg/api"
 	m "github.com/blackducksoftware/perceptor/pkg/core/model"
+	"github.com/blackducksoftware/perceptor/pkg/hub"
+	log "github.com/sirupsen/logrus"
 )
 
 // GetScanResults .....
@@ -42,4 +46,56 @@ func (g *GetScanResults) Apply(model *m.Model) {
 	go func() {
 		g.Done <- scanResults
 	}()
+}
+
+func scanResultsForPod(model *m.Model, podName string) (*m.Scan, error) {
+	pod, ok := model.Pods[podName]
+	if !ok {
+		return nil, fmt.Errorf("could not find pod of name %s in cache", podName)
+	}
+
+	overallStatus := hub.PolicyStatusTypeNotInViolation
+	policyViolationCount := 0
+	vulnerabilityCount := 0
+	for _, container := range pod.Containers {
+		imageScan, err := scanResultsForImage(model, container.Image.Sha)
+		if err != nil {
+			log.Errorf("unable to get scan results for image %s: %s", container.Image.Sha, err.Error())
+			return nil, err
+		}
+		if imageScan == nil {
+			return nil, nil
+		}
+		policyViolationCount += imageScan.PolicyViolations
+		vulnerabilityCount += imageScan.Vulnerabilities
+		imageScanOverallStatus := imageScan.OverallStatus
+		if imageScanOverallStatus != hub.PolicyStatusTypeNotInViolation {
+			overallStatus = imageScanOverallStatus
+		}
+	}
+	podScan := &m.Scan{
+		OverallStatus:    overallStatus,
+		PolicyViolations: policyViolationCount,
+		Vulnerabilities:  vulnerabilityCount}
+	return podScan, nil
+}
+
+func scanResultsForImage(model *m.Model, sha m.DockerImageSha) (*m.Scan, error) {
+	imageInfo, ok := model.Images[sha]
+	if !ok {
+		return nil, fmt.Errorf("could not find image of sha %s in cache", sha)
+	}
+
+	if imageInfo.ScanStatus != m.ScanStatusComplete {
+		return nil, nil
+	}
+	if imageInfo.ScanResults == nil {
+		return nil, fmt.Errorf("model inconsistency: could not find scan results for completed image %s", sha)
+	}
+
+	imageScan := &m.Scan{
+		OverallStatus:    imageInfo.ScanResults.OverallStatus(),
+		PolicyViolations: imageInfo.ScanResults.PolicyViolationCount(),
+		Vulnerabilities:  imageInfo.ScanResults.VulnerabilityCount()}
+	return imageScan, nil
 }
